+++ doapi/__init__.py
@@ -2,7 +2,7 @@ __version__ = '0.2.0.dev1'
 
 from .base        import (DOEncoder, Resource, Region, Size, Account, Kernel,
                           DropletUpgrade, Networks, NetworkInterface,
-                          BackupWindow, DOAPIError)
+                          BackupWindow, DOAPIError, WaitTimeoutError)
 from .action      import Action, ActionError
 from .doapi       import doapi
 from .domain      import Domain, DomainRecord